refactor(components): refactor skeleton/skeleton-item
