+++ packages/components/skeleton/index.ts
@@ -1,7 +1,7 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils/with-install'
 
-import Skeleton from './src/index.vue'
-import SkeletonItem from './src/item.vue'
+import Skeleton from './src/skeleton.vue'
+import SkeletonItem from './src/skeleton-item.vue'
 
 export const ElSkeleton = withInstall(Skeleton, {
   SkeletonItem,
@@ -9,4 +9,5 @@ export const ElSkeleton = withInstall(Skeleton, {
 export default ElSkeleton
 export const ElSkeletonItem = withNoopInstall(SkeletonItem)
 
-export * from './src/types'
+export * from './src/skeleton'
+export * from './src/skeleton-item'
