--- packages/components/skeleton/src/types.ts
@@ -1,10 +1,0 @@
-export type Variants =
-  | 'circle'
-  | 'rect'
-  | 'h1'
-  | 'h3'
-  | 'text'
-  | 'caption'
-  | 'p'
-  | 'image'
-  | 'button'
